live/requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
cryptography>=42.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
